Updated clustering
Removed the previous "convex hull" method for generating the area of a cluster and moved to finite radius circles around each point in the circle.

Added method to RelateObjects to use children of an object.  In the present case these are the spots that make up a cluster.  These are used up to the same epsilon range as in DBSCAN.

Added SaveImages option to store output images in a directory structure mirroring the input structure.  This will allow us to store images on a separate HDD; hopefully increasing performance when IO-limited.

+++ src/main/java/wbif/sjx/ModularImageAnalysis/Object/ObjSet.java
@@ -28,11 +28,11 @@ public class ObjSet extends LinkedHashMap<Integer,Obj> {
         return maxID;
     }
 
-    public double[][] getSpatialLimits() {
-        double[][] limits = new double[][]{
-                {Double.MAX_VALUE,Double.MIN_VALUE},
-                {Double.MAX_VALUE,Double.MIN_VALUE},
-                {Double.MAX_VALUE,Double.MIN_VALUE}};
+    public int[][] getSpatialLimits() {
+        int[][] limits = new int[][]{
+                {Integer.MAX_VALUE,Integer.MIN_VALUE},
+                {Integer.MAX_VALUE,Integer.MIN_VALUE},
+                {Integer.MAX_VALUE,Integer.MIN_VALUE}};
 
         for (Obj object:values()) {
             ArrayList<Integer> x = object.getCoordinates(Obj.X);
@@ -53,4 +53,5 @@ public class ObjSet extends LinkedHashMap<Integer,Obj> {
         return limits;
 
     }
+
 }
